backend/requirements.txt
Django==5.2.4
openai==1.98.0
selenium==4.34.2
requests==2.32.4
beautifulsoup4==4.13.4
python-dotenv==1.1.1
gunicorn==23.0.0
webdriver-manager==4.0.2